Moving parser to separate package
