+++ app/src/main/java/acr/browser/lightning/adblock/source/AssetsHostsDataSource.kt
@@ -1,7 +1,7 @@
 package acr.browser.lightning.adblock.source
 
 import acr.browser.lightning.BuildConfig
-import acr.browser.lightning.adblock.HostsFileParser
+import acr.browser.lightning.adblock.parser.HostsFileParser
 import acr.browser.lightning.log.Logger
 import android.content.res.AssetManager
 import io.reactivex.Single
